Implemented checkCheck
Created the checkCheck function which checks whether or not a given player has the other in check or not. Also fixed a bug in checkIntermediatePieces where test_pos was created via pass by reference, which messed up the board state. Created a few helper/diagnostic functions to help with checkCheck and future possible functions.